chore: rename repos module to app-store
