+++ packages/backend/src/modules/app-stores/dto/app-store.dto.ts
@@ -1,0 +1,10 @@
+import { createZodDto } from 'nestjs-zod';
+import { z } from 'zod';
+
+// Pull
+export class PullDto extends createZodDto(
+  z.object({
+    success: z.boolean(),
+  }),
+) {}
+
