+++ packages/backend/src/modules/docker/docker.module.ts
@@ -1,11 +1,11 @@
 import { SocketModule } from '@/core/socket/socket.module';
 import { Module } from '@nestjs/common';
+import { AppStoreModule } from '../app-stores/app-store.module';
 import { AppsModule } from '../apps/apps.module';
-import { ReposModule } from '../repos/repos.module';
 import { DockerService } from './docker.service';
 
 @Module({
-  imports: [AppsModule, ReposModule, SocketModule],
+  imports: [AppsModule, AppStoreModule, SocketModule],
   providers: [DockerService],
   exports: [DockerService],
 })
