+++ packages/backend/src/modules/docker/docker.service.ts
@@ -6,9 +6,9 @@ import { FilesystemService } from '@/core/filesystem/filesystem.service';
 import { LoggerService } from '@/core/logger/logger.service';
 import { SocketManager } from '@/core/socket/socket.service';
 import { Injectable } from '@nestjs/common';
+import { ReposHelpers } from '../app-stores/repos.helpers';
 import { AppFilesManager } from '../apps/app-files-manager';
 import type { AppEventFormInput } from '../queue/entities/app-events';
-import { ReposHelpers } from '../repos/repos.helpers';
 import { DockerComposeBuilder } from './builders/compose.builder';
 import { type Service, type ServiceInput, serviceSchema } from './builders/schemas';
 import { ServiceBuilder } from './builders/service.builder';
