--- packages/backend/src/modules/repos/dto/repos.dto.ts
@@ -1,7 +1,0 @@
-import { createZodDto } from "nestjs-zod";
-import { z } from "zod";
-
-// Pull
-export class PullDto extends createZodDto(z.object({
-    success: z.boolean(),
-})) {}
