+++ packages/backend/src/swagger.json
@@ -643,7 +643,7 @@
           }
         },
         "tags": [
-          "Repos"
+          "AppStore"
         ]
       }
     },
